New folder/requirements.txt
Flask
Pillow
reportlab
werkzeug